Many changes -- small cleanups and moving to AutoHost.
there are lots of little crash fixes. significant changes
are in _initialize.lua (which is due for a rename). rather than
talking with a widget, it talks with a SPADS plugin to load/save data.

this allows armies to be saved across games run on a spring-dedicated
based autohost. this is good because many of them will happily run on a
very low-power server.